[Fix] 파일 및 해시값 json 생성 API 클래스화 및 해당 클래스 호출 함수 agent-update의 index.js에 추가
